My_Bot/requirements.txt
# --- FastAPI e ASGI ---
anyio==4.9.0
fastapi==0.115.12
h11==0.16.0
httptools==0.6.4
sniffio==1.3.1
starlette==0.46.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1

# --- Telegram Bot ---
python-telegram-bot==20.7

# --- Validação e Tipagem ---
annotated-types==0.7.0
pydantic==2.11.4
pydantic_core==2.33.2
typing-inspection==0.4.0
typing_extensions==4.13.2

# --- Web Scraping e HTTP ---
aiohttp==3.9.3
beautifulsoup4==4.12.3
bs4==0.0.1  # Alias para beautifulsoup4
requests==2.31.0

# --- Utilitários e Outras Dependências ---
click==8.1.8
colorama==0.4.6
idna==3.10
python-dotenv==1.1.0
PyYAML==6.0.2
